requirements.txt
telethon==1.40.0
python-dotenv==1.1.1
transformers==4.55.2
torch==2.8.0
spacy==3.8.7
pandas==2.3.1
scikit-learn==1.7.1
nltk==3.9.1
textblob==0.19.0
geopy==2.4.1
protobuf==6.32.0
sentencepiece==0.2.1
python-telegram-bot==22.3